moved getAddableComponents to separate service
