+++ services/components/add-component-handler.test.js
@@ -1,8 +1,7 @@
 var dirname = __dirname.split('/').pop(),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require('./add-component-handler'),
-  dom = require('@nymag/dom'),
-  site = require('../site');
+  dom = require('@nymag/dom');
 
 describe(dirname, function () {
   describe(filename, function () {
@@ -40,54 +39,6 @@ describe(dirname, function () {
         var el = dom.create('<div><div class="inner" data-editable="content"></div></div>');
 
         expect(fn(el, 'content')).to.equal(el.querySelector('.inner'));
-      });
-    });
-
-    describe('getAddableComponents', function () {
-      var fn = lib[this.title],
-        sandbox;
-
-      beforeEach(function () {
-        sandbox = sinon.sandbox.create();
-        sandbox.stub(site, 'get').returns('testSite');
-      });
-
-      afterEach(function () {
-        sandbox.restore();
-      });
-
-      it('works for empty array', function () {
-        expect(fn([])).to.deep.equal([]);
-      });
-
-      it('works when excluding nothing', function () {
-        expect(fn(['foo'])).to.deep.equal(['foo']);
-      });
-
-      it('works when excluding empty array', function () {
-        expect(fn(['foo'], [])).to.deep.equal(['foo']);
-      });
-
-      it('works when excluding stuff', function () {
-        expect(fn(['foo'], ['bar'])).to.deep.equal(['foo']);
-      });
-
-      it('allows components included on current site', function () {
-        expect(fn(['foo (testSite)'])).to.deep.equal(['foo']);
-        expect(fn(['foo (testSite, otherSite)'])).to.deep.equal(['foo']);
-      });
-
-      it('disallows components not included on current site', function () {
-        expect(fn(['foo (otherSite)'])).to.deep.equal([]);
-      });
-
-      it('disallows components excluded from current site', function () {
-        expect(fn(['foo (not:testSite)'])).to.deep.equal([]);
-        expect(fn(['foo (not: testSite)'])).to.deep.equal([]);
-      });
-
-      it('disallows components included but also excluded from current site', function () {
-        expect(fn(['foo (testSite, otherSite, not:testSite)'])).to.deep.equal([]);
       });
     });
   });
